Animation fix and winmm & dinput stuff

+++ ShinRyuModManager/Program.cs
@@ -421,59 +421,64 @@ namespace ShinRyuModManager
                 }
             }
 
-            // TODO: Maybe move this to a separate "Game patches" file
-            // Virtua Fighter eSports crashes when used with dinput8.dll as the ASI loader
-            if (game == Game.eve && File.Exists(DINPUT8DLL))
+
+            if(File.Exists(DINPUT8DLL) && File.Exists(WINMMLJ))
             {
-                if (File.Exists(VERSIONDLL))
+                // TODO: Maybe move this to a separate "Game patches" file
+                // Virtua Fighter eSports crashes when used with dinput8.dll as the ASI loader
+                if (game == Game.eve && File.Exists(DINPUT8DLL))
                 {
-                    Console.Write($"Game specific patch: Deleting {DINPUT8DLL} because {VERSIONDLL} exists...");
-
-                    // Remove dinput8.dll
-                    File.Delete(DINPUT8DLL);
-                }
-                else
-                {
-                    Console.Write($"Game specific patch: Renaming {DINPUT8DLL} to {VERSIONDLL}...");
-
-                    // Rename dinput8.dll to version.dll to prevent the game from crashing
-                    File.Move(DINPUT8DLL, VERSIONDLL);
-                }
-
-                Console.WriteLine(" DONE!\n");
-            }
-            else if (game >= Game.YakuzaKiwami2 && game != Game.likeadragongaiden)
-            {
-                //
-                // Lost Judgment (and Judgment post update 1) does not like Ultimate ASI Loader, so instead we use a custom build of DllSpoofer (https://github.com/Kazurin-775/DllSpoofer)
-                if (File.Exists(DINPUT8DLL))
-                {
-                    Console.Write($"Game specific patch: Deleting {DINPUT8DLL} because it causes crashes with Judgment games...");
-
-                    // Remove dinput8.dll
-                    File.Delete(DINPUT8DLL);
-
-                    Console.WriteLine(" DONE!\n");
-                }
-
-                if (!File.Exists(WINMMDLL))
-                {
-                    if (File.Exists(WINMMLJ))
+                    if (File.Exists(VERSIONDLL))
                     {
-                        Console.Write($"Game specific patch: Enabling {WINMMDLL} by renaming {WINMMLJ} to fix Judgment games crashes...");
+                        Console.Write($"Game specific patch: Deleting {DINPUT8DLL} because {VERSIONDLL} exists...");
 
-                        // Rename dinput8.dll to version.dll to prevent the game from crashing
-                        File.Move(WINMMLJ, WINMMDLL);
-
-                        Console.WriteLine(" DONE!\n");
+                        // Remove dinput8.dll
+                        File.Delete(DINPUT8DLL);
                     }
                     else
                     {
-                        Console.ForegroundColor = ConsoleColor.Red;
-                        Console.WriteLine($"WARNING: {WINMMLJ} was not found. Judgment games will NOT load mods without this file. Please redownload Shin Ryu Mod Manager.\n");
-                        Console.ResetColor();
+                        Console.Write($"Game s  pecific patch: Renaming {DINPUT8DLL} to {VERSIONDLL}...");
+
+                        // Rename dinput8.dll to version.dll to prevent the game from crashing
+                        File.Move(DINPUT8DLL, VERSIONDLL);
+                    }
+
+                    Console.WriteLine(" DONE!\n");
+                }
+                else if (game >= Game.YakuzaKiwami2 && game != Game.likeadragongaiden)
+                {
+                    //
+                    // Lost Judgment (and Judgment post update 1) does not like Ultimate ASI Loader, so instead we use a custom build of DllSpoofer (https://github.com/Kazurin-775/DllSpoofer)
+                    if (File.Exists(DINPUT8DLL))
+                    {
+                        Console.Write($"Game specific patch: Deleting {DINPUT8DLL} because it causes crashes with Judgment games...");
+
+                        // Remove dinput8.dll
+                        File.Delete(DINPUT8DLL);
+
+                        Console.WriteLine(" DONE!\n");
+                    }
+
+                    if (!File.Exists(WINMMDLL))
+                    {
+                        if (File.Exists(WINMMLJ))
+                        {
+                            Console.Write($"Game specific patch: Enabling {WINMMDLL} by renaming {WINMMLJ} to fix Judgment games crashes...");
+
+                            // Rename dinput8.dll to version.dll to prevent the game from crashing
+                            File.Move(WINMMLJ, WINMMDLL);
+
+                            Console.WriteLine(" DONE!\n");
+                        }
+                        else
+                        {
+                            Console.ForegroundColor = ConsoleColor.Red;
+                            Console.WriteLine($"WARNING: {WINMMLJ} was not found. Judgment games will NOT load mods without this file. Please redownload Shin Ryu Mod Manager.\n");
+                            Console.ResetColor();
+                        }
                     }
                 }
+
             }
 
             // Read ini (again) to check if we should try importing the old load order file
